Move the multimodule harnessing to the only place where it is used.
