+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/MultiModuleProjectTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2021 DiffPlug
+ * Copyright 2016-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,15 @@
  */
 package com.diffplug.spotless.maven;
 
-import static com.diffplug.spotless.maven.MavenIntegrationHarness.SubProjectFile.file;
+import static java.util.Arrays.asList;
+import static java.util.Collections.emptyMap;
+import static java.util.Collections.singletonMap;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 import org.junit.jupiter.api.Test;
 
@@ -95,5 +103,92 @@ class MultiModuleProjectTest extends MavenIntegrationHarness {
 
 		assertFile("three/src/main/scala/test1.scala").sameAsResource("scala/scalafmt/basic.cleanWithCustomConf_3.0.0");
 		assertFile("three/src/test/scala/test2.scala").sameAsResource("scala/scalafmt/basic.cleanWithCustomConf_3.0.0");
+	}
+
+	private static final String CHILD_ID = "childId";
+
+	protected MultiModuleProjectCreator multiModuleProject() {
+		return new MultiModuleProjectCreator();
+	}
+
+	class MultiModuleProjectCreator {
+		private String configSubProject;
+		private SubProjectFile[] configSubProjectFiles;
+		private String[] configuration;
+		private final Map<String, List<SubProjectFile>> subProjects = new LinkedHashMap<>();
+
+		protected MultiModuleProjectCreator withConfigSubProject(String name, SubProjectFile... files) {
+			configSubProject = name;
+			configSubProjectFiles = files;
+			return this;
+		}
+
+		protected MultiModuleProjectCreator withConfiguration(String... lines) {
+			configuration = lines;
+			return this;
+		}
+
+		protected MultiModuleProjectCreator addSubProject(String name, SubProjectFile... files) {
+			subProjects.put(name, asList(files));
+			return this;
+		}
+
+		protected void create() throws IOException {
+			createRootPom();
+			createConfigSubProject();
+			createSubProjects();
+		}
+
+		private void createRootPom() throws IOException {
+			List<String> modulesList = new ArrayList<>();
+			modulesList.add(configSubProject);
+			modulesList.addAll(subProjects.keySet());
+			String[] modules = modulesList.toArray(new String[0]);
+
+			Map<String, Object> rootPomParams = buildPomXmlParams(null, null, configuration, modules, null);
+			setFile("pom.xml").toContent(createPomXmlContent("/multi-module/pom-parent.xml.mustache", rootPomParams));
+		}
+
+		private void createConfigSubProject() throws IOException {
+			if (configSubProject != null) {
+				String content = createPomXmlContent("/multi-module/pom-config.xml.mustache", emptyMap());
+				setFile(configSubProject + "/pom.xml").toContent(content);
+
+				createSubProjectFiles(configSubProject, asList(configSubProjectFiles));
+			}
+		}
+
+		private void createSubProjects() throws IOException {
+			for (Map.Entry<String, List<SubProjectFile>> entry : subProjects.entrySet()) {
+				String subProjectName = entry.getKey();
+				List<SubProjectFile> subProjectFiles = entry.getValue();
+
+				String content = createPomXmlContent("/multi-module/pom-child.xml.mustache", singletonMap(CHILD_ID, subProjectName));
+				setFile(subProjectName + "/pom.xml").toContent(content);
+
+				createSubProjectFiles(subProjectName, subProjectFiles);
+			}
+		}
+
+		private void createSubProjectFiles(String subProjectName, List<SubProjectFile> subProjectFiles) {
+			for (SubProjectFile file : subProjectFiles) {
+				setFile(subProjectName + '/' + file.to).toResource(file.from);
+			}
+		}
+	}
+
+	static class SubProjectFile {
+
+		private final String from;
+		private final String to;
+
+		private SubProjectFile(String from, String to) {
+			this.from = from;
+			this.to = to;
+		}
+	}
+
+	static SubProjectFile file(String from, String to) {
+		return new SubProjectFile(from, to);
 	}
 }
